Added dark mode, language switching, and a new settings page for admins
  - Cleaned up main.dart so it only does the basics (Firebase setup + launching the app)
  - The admin dashboard now keeps track of dark mode and language on its own and passes them down to the settings page
  - Built a new admin settings page where you can turn dark mode on/off, switch between English and French, toggle maintenance mode, and update the support email, all
  saved to the database
  - Fixed a bug on the login screen where email verification wasn't always up to date
  - Added a proper error message when bookings fail to load
  - Fixed a crash caused by a price field being the wrong number type in a few screens

+++ lib/screens/owner/booking_screen.dart
@@ -8,7 +8,7 @@ import '../../models/pet_model.dart';
 class BookingScreen extends StatefulWidget {
   final String walkerId;
   final String walkerName;
-  final int hourlyRate;
+  final double hourlyRate;
 
   const BookingScreen({
     super.key,
